Consolidar dados Relatório

+++ src/SME.SR.Infra/Dtos/Alunos/AlunoReduzidoDto.cs
@@ -2,8 +2,8 @@
 {
     public class AlunoReduzidoDto
     {
-        public int TurmaCodigo { get; set; }
-        public int AlunoCodigo { get; set; }
+        public long TurmaCodigo { get; set; }
+        public long AlunoCodigo { get; set; }
         public string NomeAluno { get; set; }
         public string NomeSocialAluno { get; set; }
